requirements.txt
# LLM
openai==1.14.2
python-dotenv==1.0.0
httpx==0.23.0

# Tools
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.20.0
pandas==2.2.1

# Google APIs
google-api-python-client==2.126.0
googlemaps==4.10.0
gspread==6.0.0
oauth2client==4.1.3

# Perplexity
perplexipy==1.3.1
